pp/sr: extract json schema dialect for reuse
Want to use this in Iceberg JSON Schema integration for consistent
behavior between SR and other Redpanda components interacting with
schemas via SR.
