+++ src/v/pandaproxy/BUILD
@@ -111,6 +111,7 @@ redpanda_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         ":config",
+        ":json_schema_dialect",
         ":rest_swagger",
         ":schema_registry_swagger",
         ":subject_name_strategy",
@@ -214,6 +215,18 @@ redpanda_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//src/v/base",
+        "//src/v/strings:string_switch",
+    ],
+)
+
+redpanda_cc_library(
+    name = "json_schema_dialect",
+    hdrs = [
+        "schema_registry/json_schema/dialect.h",
+    ],
+    include_prefix = "pandaproxy",
+    visibility = ["//visibility:public"],
+    deps = [
         "//src/v/strings:string_switch",
     ],
 )
